fix Photo

+++ src/photos/photos.controller.ts
@@ -1,5 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseInterceptors, ClassSerializerInterceptor, UseGuards, UploadedFile, BadRequestException, NotFoundException } from '@nestjs/common';
-import { PhotosService } from './photos.service';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseInterceptors, ClassSerializerInterceptor, UseGuards, UploadedFile, NotFoundException, Res } from '@nestjs/common';
 import { CreatePhotoDto } from './dto/create-photo.dto';
 import { UpdatePhotoDto } from './dto/update-photo.dto';
 import { ApiBody, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
@@ -10,18 +9,21 @@ import { FileInterceptor } from '@nestjs/platform-express/multer';
 import { imageFileFilter } from './middleware/imageFileFilter';
 import { diskStorage } from 'multer';
 import { extname } from 'path';
-
+import { GetUser } from 'src/auth/get-user.decorator';
+import PhotosService from './photos.service';
 
 
 
 @ApiTags('PHOTOS') // Create a PHOTOS category in swagger UI.
 @Controller('photos')
-export class PhotosController
+export default class PhotosController
 {
   constructor(
     private readonly photosService: PhotosService,
     private readonly userService: UsersService,
-    private readonly albumService: AlbumsService) { }
+    private readonly albumService: AlbumsService) {}
+
+
 
   @UseGuards(JwtAuthGuard) // The user must be logged in / registered.
   @UseInterceptors(ClassSerializerInterceptor) // Does not return entity properties marked with @Exclude()
@@ -42,10 +44,20 @@ export class PhotosController
     fileFilter: imageFileFilter,
   }))
   @Post('uploads')
-  async uploadFile(@UploadedFile() file: Express.Multer.File)
+  async uploadFile(
+    @Body() createPhotoDto: CreatePhotoDto,
+    @GetUser() user,
+    @UploadedFile() savedFiles: Array<Express.Multer.File>)
   {
-    console.log(file);
-    const newPhoto = await this.photosService.create(file)
+    console.log(savedFiles);
+    const oneUser = await this.userService.findOne(user.userId);
+    const albumExist = await this.albumService.findAlbumById(createPhotoDto.albumId);
+    if (!albumExist) throw new NotFoundException(`The photo album does not exist.`);
+    const newPhoto = await this.photosService.create(
+      oneUser,
+      savedFiles,
+      createPhotoDto
+    )
     return {
       statusCode: 201,
       message: `Successful creation of a new upload`,
@@ -59,27 +71,19 @@ export class PhotosController
 
   @UseGuards(JwtAuthGuard) // The user must be logged in / registered.
   @UseInterceptors(ClassSerializerInterceptor) // Does not return entity properties marked with @Exclude()
-  @Get()
-async findAllPhotos(){
-  const photos = await this.photosService.findAll();
-  if (!photos)
+  @Get(':imgpath')
+  async findAllPhotos (@Param('imgpath') file, @Res() res)
   {
-    throw new NotFoundException(`Albums not found.`);
+    return res.sendFile(file, { root: '.uploads' });
   }
-  return {
-    statusCode: 200,
-    message: `List of all albums.`,
-    data: photos
-  };
-}
 
 
 
 
 
-  @UseGuards(JwtAuthGuard) // The user must be logged in / registered.
-  @UseInterceptors(ClassSerializerInterceptor) // Does not return entity properties marked with @Exclude()
-  @Get(':id')
+  //@UseGuards(JwtAuthGuard) // The user must be logged in / registered.
+  //@UseInterceptors(ClassSerializerInterceptor) // Does not return entity properties marked with @Exclude()
+  /* @Get(':id')
   async findOnePhoto(@Param('id') id: number)
   {
     const onePhoto = await this.photosService.findOne(id);
@@ -92,7 +96,7 @@ async findAllPhotos(){
       message: `Photo with id ${id} is found.`,
       data: onePhoto
     }
-  }
+  } */
 
 
 
@@ -116,20 +120,21 @@ async findAllPhotos(){
     fileFilter: imageFileFilter,
   }))
   @Patch(':id')
-  async update(@Param('id') id: number, @Body() updatePhotoDto: UpdatePhotoDto, @UploadedFile() file: Express.Multer.File)
+  async update(@Param('id') id: number, @Body() updatePhotoDto: UpdatePhotoDto, /* @UploadedFile() file: Express.Multer.File */)
   {
     const photoExist = await this.photosService.findOne(id);
     if (!photoExist)
     {
-      throw new BadRequestException(`Photo not found`);
+      throw new NotFoundException(`Photo not found`);
     }
+    const albumExist = await this.albumService.findAlbumById(updatePhotoDto.albumId);
 
-    const photo = await this.photosService.update(id, updatePhotoDto, file);
+    const photoUpdated = await this.photosService.update(id, updatePhotoDto/* , file */);
 
     return {
       statusCode: 200,
       message: `Success of change.`,
-      data: photo
+      data: photoUpdated
     };
 
   }
@@ -140,9 +145,10 @@ async findAllPhotos(){
   @Delete(':id')
   async remove(@Param('id') id: number)
   {
+    const photoExist = await this.photosService.findOne(id);
+    if (!photoExist) throw new NotFoundException(`This photo does not exist.`);
+
     const deletedPhoto = await this.photosService.remove(id);
-
-
     return {
       statusCode: 200,
       message: `Deleted photo.`,
